+++ packages/DataTable/src/DataTable.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { useTable, useBlockLayout, Column, Row } from "react-table";
-import ResizeDetector from "@paprika/resize-detector";
 import { extractChildren } from "@paprika/helpers";
 import { gridTypes } from "@paprika/constants";
 import { InfiniteLoader, InfiniteLoaderImpl, InfiniteLoaderPublicProps } from "./components/InfiniteLoader";
@@ -11,14 +10,10 @@ import { TableHeader } from "./components/TableHeader";
 import { TableBody } from "./components/TableBody";
 import { TableRow } from "./components/TableRow";
 import useSticky from "./hooks/useSticky";
-import useTableDimensions from "./hooks/useTableDimensions";
-
+import useBestTableDimensions from "./hooks/useBestTableDimensions";
 import { TableDataItemType, BorderType } from "./types";
 
 import * as sc from "./DataTable.styles";
-
-export const ResizeContainer: React.FC<Record<string, any>> = (...moreProps) => <></>;
-ResizeContainer.displayName = "DataTable.ResizeContainer";
 
 function InnerElement({ children, ...rest }: { children: React.ReactNode }): JSX.Element {
   return (
@@ -35,7 +30,6 @@ interface ConstantsTypes {
 
 interface DataTableComposition {
   InfiniteLoader: React.FC<InfiniteLoaderPublicProps>;
-  ResizeContainer: React.FC<Record<string, any>>;
   types: ConstantsTypes;
 }
 
@@ -51,11 +45,11 @@ export interface DataTableProps {
   /** Array of data to be stored in the table */
   data: TableDataItemType[];
 
-  /** The height of the table in px */
-  height?: number;
+  /** The max height of the table in px/vh/calc(xvh+/-xpx) */
+  maxHeight?: string;
 
-  /** The width of the table */
-  width?: string | number;
+  /** The max width of the table in px/vw/calc(xvw+/-xpx) */
+  maxWidth?: string;
 
   /** Define the look for borders in the table */
   borderType?: BorderType;
@@ -72,29 +66,42 @@ export interface DataTableProps {
   /** Override the table Column configuration for some particular rows */
   renderRow?: ((row: Row<TableDataItemType>) => React.ReactNode) | null;
 
+  /** If the table should be resized when the viewport size changed */
+  shouldResizeWithViewport?: boolean;
+
   /** Experimental prop */
   extraCellProps?: Record<string, any>;
 
   [x: string]: any;
 }
 
-const DataTable: React.FC<DataTableProps> & DataTableComposition = ({
-  a11yText,
-  children = null,
-  columns,
-  data,
-  height,
-  width,
-  borderType = gridTypes.HORIZONTAL,
-  getRowHeight = null,
-  hasZebraStripes = false,
-  isHeaderSticky = true,
-  renderRow = null,
-  extraCellProps = {},
-  ...moreProps
-}: DataTableProps) => {
+function DataTable(
+  {
+    a11yText,
+    children = null,
+    columns,
+    data,
+    maxHeight = "100vh",
+    maxWidth = "100vw",
+    borderType = gridTypes.HORIZONTAL,
+    getRowHeight = null,
+    hasZebraStripes = false,
+    isHeaderSticky = true,
+    renderRow = null,
+    shouldResizeWithViewport = true,
+    extraCellProps = {},
+    ...moreProps
+  }: DataTableProps,
+  componentRef: React.RefObject<unknown>
+) {
   const tableRef = React.useRef<HTMLDivElement>(null);
-  const { getTableHeight, getTableWidth } = useTableDimensions(tableRef, data.length, height, width);
+  const bestDimensions = useBestTableDimensions({
+    tableRef,
+    componentRef,
+    maxHeight,
+    maxWidth,
+    shouldResizeWithViewport,
+  });
 
   const defaultColumn = React.useMemo(
     () => ({
@@ -118,12 +125,10 @@ const DataTable: React.FC<DataTableProps> & DataTableComposition = ({
   const {
     // @ts-ignore
     "DataTable.InfiniteLoader": extractedInfiniteLoaderDefinition,
-    // @ts-ignore
-    "DataTable.ResizeContainer": extractedResizeContainer,
-  } = extractChildren(children, ["DataTable.InfiniteLoader", "DataTable.ResizeContainer"]);
+  } = extractChildren(children, ["DataTable.InfiniteLoader"]);
   /* eslint-enable @typescript-eslint/ban-ts-comment */
 
-  function renderTableContent(maxHeight: number) {
+  function renderTableContent() {
     const hasInfiniteLoader = Boolean(extractedInfiniteLoaderDefinition);
 
     if (hasInfiniteLoader) {
@@ -132,9 +137,10 @@ const DataTable: React.FC<DataTableProps> & DataTableComposition = ({
       return (
         <InfiniteLoaderImpl
           data={data}
-          height={maxHeight}
+          height={bestDimensions.height}
           innerElementType={InnerElement}
           getRowHeight={getRowHeight}
+          shouldHaveHorizontalScroll={bestDimensions.shouldHaveHorizontalScroll}
           {...infiniteLoaderPublicProps}
         />
       );
@@ -149,59 +155,40 @@ const DataTable: React.FC<DataTableProps> & DataTableComposition = ({
     );
   }
 
-  function renderTable(height: number, width: number | string) {
-    return (
+  return (
+    <ThemeContext.Provider value={{ borderType, isHeaderSticky, hasZebraStripes }}>
       <sc.Table
         ref={tableRef}
         aria-label={a11yText}
         aria-rowcount={data.length}
         data-pka-anchor="dataTable"
-        width={width}
-        height={height}
+        width={bestDimensions.width}
+        height={bestDimensions.height}
         {...tableInstance.getTableProps()}
         {...moreProps}
       >
         <sc.ContentWrapper>
           <RenderRowContext.Provider value={renderRow}>
-            <ReactTableContext.Provider value={tableInstance}>{renderTableContent(height)}</ReactTableContext.Provider>
+            <ReactTableContext.Provider value={tableInstance}>{renderTableContent()}</ReactTableContext.Provider>
           </RenderRowContext.Provider>
         </sc.ContentWrapper>
       </sc.Table>
-    );
-  }
-
-  if (!extractedResizeContainer) {
-    if (width === undefined) {
-      console.warn("The `width` prop should be specified if `<DataTable.ResizeContainer>` is not used");
-    }
-
-    if (height === undefined) {
-      console.warn("The `height` prop should be specified if `<DataTable.ResizeContainer>` is not used");
-    }
-  }
-
-  return (
-    <ThemeContext.Provider value={{ borderType, isHeaderSticky, hasZebraStripes }}>
-      {extractedResizeContainer ? (
-        <div {...extractedResizeContainer.props}>
-          <ResizeDetector isFullHeight>
-            {({ height: maxHeight, width: maxWidth }) =>
-              renderTable(getTableHeight(maxHeight), getTableWidth(maxWidth))
-            }
-          </ResizeDetector>
-        </div>
-      ) : (
-        renderTable(height || 0, width || 0)
-      )}
     </ThemeContext.Provider>
   );
-};
+}
 
-DataTable.displayName = "DataTable";
-DataTable.InfiniteLoader = InfiniteLoader;
-DataTable.ResizeContainer = ResizeContainer;
-DataTable.types = {
+declare module "react" {
+  function forwardRef<T, P = unknown>(
+    render: (props: P, ref: React.RefObject<T>) => React.ReactElement | null
+  ): React.FC<DataTableProps> & DataTableComposition;
+}
+
+const DataTableWithRef = React.forwardRef(DataTable);
+
+DataTableWithRef.displayName = "DataTable";
+DataTableWithRef.InfiniteLoader = InfiniteLoader;
+DataTableWithRef.types = {
   borderType: gridTypes,
 };
 
-export default DataTable;
+export default DataTableWithRef;