+++ packages/DataTable/stories/examples/RealWorld.tsx
@@ -213,6 +213,7 @@ export const RealWorldStory: (props: { isForTesting?: boolean }) => JSX.Element 
 
       <DataTable
         a11yText="Data table for a real world example."
+        maxHeight="calc(100vh-200px)"
         columns={columns}
         data={data}
         borderType={borderType}
@@ -224,7 +225,6 @@ export const RealWorldStory: (props: { isForTesting?: boolean }) => JSX.Element 
           isNextPageLoading={false}
           loadMoreItems={handleLoadMore}
         />
-        <DataTable.ResizeContainer style={{ height: "calc(100vh - 200px)", width: "100%" }} />
       </DataTable>
     </>
   );